requirements.txt
opencv-python==4.10.0.84
pyautogui==0.9.54
pillow==10.1.0
tkcap==0.0.3
pydicom==2.4.4
img2pdf==0.5.1
matplotlib==3.9.1
pandas==2.2.2
tensorflow==2.15.0
python-xlib==0.33
tk==0.1.0
easy-pil==0.3.3
numpy==1.26.4
keras==2.15.0
torch==2.1.0
opencv-python==4.10.0.84
streamlit==1.37.1
gdown==5.2.0
fpdf==1.7.2